[MOL-10990][ZX|KT] correction
