+++ src/input-textarea/textarea.tsx
@@ -76,7 +76,7 @@ const TextareaComponent = (
                     disabled={disabled}
                     value={stateValue}
                     maxLength={otherProps.maxLength}
-                    counterLabelFunction={otherProps.counterLabelFunction}
+                    renderCustomCounter={otherProps.renderCustomCounter}
                 />
             )}
         </Wrapper>
