+++ stories/form/form-textarea/form-textarea.stories.mdx
@@ -10,7 +10,8 @@ import {
     Title,
 } from "../../storybook-common";
 import { Container } from "../shared-doc-elements";
-import { PropsTable, counterLabelFunction } from "./props-table";
+import { PropsTable } from "./props-table";
+import { renderCustomCounter } from "./doc-elements.tsx";
 
 <Meta title="Form/Textarea" component={Form.Textarea} />
 
@@ -53,13 +54,6 @@ import { Form } from "@lifesg/react-design-system/form";
                     label="This is does not have a counter"
                     placeholder="Enter here..."
                 />
-                <br />
-                <Form.Textarea
-                    label="This is a textarea with a custom counter"
-                    placeholder="Enter here..."
-                    maxLength={100}
-                    counterLabelFunction={counterLabelFunction}
-                />
             </Container>
         </StoryContainer>
     </Story>
@@ -78,6 +72,23 @@ import { Textarea } from "@lifesg/react-design-system/input-textarea";
         <StoryContainer>
             <Container>
                 <Textarea placeholder="Enter here..." maxLength={100} />
+            </Container>
+        </StoryContainer>
+    </Story>
+</Preview>
+
+<Heading3>Textarea with custom counter label</Heading3>
+
+<Preview>
+    <Story name="With custom counter">
+        <StoryContainer>
+            <Container>
+                <Form.Textarea
+                    label="This is a textarea with a custom counter label"
+                    placeholder="Enter here..."
+                    maxLength={100}
+                    renderCustomCounter={renderCustomCounter}
+                />
             </Container>
         </StoryContainer>
     </Story>
